requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
build==1.2.2.post1
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.2
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.7.0
django-colorfield==0.14.0
django-cors-headers==4.7.0
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
kombu==5.5.3
oauthlib==3.2.2
packaging==25.0
payme-pkg==3.0.24
pillow==11.2.1
pip-tools==7.4.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
redis==6.0.0
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==80.1.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
text-unidecode==1.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
